Changes to be committed: 	modified:   permanent_storage.cpp
The Arduino IDE build was OK, but Travis build failed.
I therefore updated type uint8_t to int as to match the types of "external int" for active_extruder and previous_extruder

+++ MM-control-01/permanent_storage.cpp
@@ -100,7 +100,7 @@ uint16_t BowdenLength::getPrevious()
 //!
 //! Returns stored value
 //! @return stored bowden length for specified extruder
-uint16_t BowdenLength::getForExtruder(uint8_t filament)
+uint16_t BowdenLength::getForExtruder(int filament)
 {
 	if (validFilament(filament))
 	{
